The migration to add a boolean admin attribute to users.
We will identify privileged administrative users with a boolean admin
attribute in the User model, which will lead automatically to an admin?
boolean method to test for admin status.

The resulting data model appears in table below:

  ### Users

 |                 |	      |
 |-----------------|----------|
 | id              | integer  |
 | name            | string   |
 | email           | string   |
 | created_at      | datetime |
 | updated_at      | datetime |
 | password_digest | string   |
 | remember_digest | string   |
 | admin 	   | boolean  |

 As usual, we add the admin attribute with a migration, indicating the
boolean type on the command line:

 ```ruby
 $ rails generate migration add_admin_to_users admin:boolean
 ```

 The migration adds the admin column to the users table, as shown in
Listing 10.54. Note that we’ve added the argument default: false to
add_column in Listing 10.54, which means that users will not be
administrators by default. (Without the default: false argument, admin
will be nil by default, which is still false, so this step is not
strictly necessary. It is more explicit, though, and communicates our
intentions more clearly both to Rails and to readers of our code.)

 Next, we migrate as usual:

$ rails db:migrate