Exercises - should not allow the admin attribute to be edited via the web
This question wasn't that straight forward to me when I first read it.

I was thinking that that we needed to pass in the password and
password_confirmation from wthin the object reference from @other_user
instance variable but in case we are concern about any attribute that
isn’t editable and writing such a test for the admin attribute

You must reload the instance variable after making changes to the
underlying record. This will load in the new changes but we must make
sure that that password relate params are set to empty string.

[should not allow the admin attribute to be edited via the web]
(https://stackoverflow.com/questions/26270102/railstutorial-org-michael-hartl-chapter-9-exercise-3-should-not-allow-the-admin/26270476)