User delete links (viewable only by admins).
The final step needed to complete the Users resource is to add delete
links and a destroy action. We’ll start by adding a delete link for each
user on the users index page, restricting access to administrative
users. The resulting "delete" links will be displayed only if the
current user is an admin

Note the method: :delete argument, which arranges for the link to issue
the necessary DELETE request. We’ve also wrapped each link inside an if
statement so that only admins can see them.

Web browsers can’t send DELETE requests natively, so Rails fakes them
with JavaScript. This means that the delete links won’t work if the user
has JavaScript disabled. If you must support non-JavaScript-enabled
browsers you can fake a DELETE request using a form and a POST request,
which works even without JavaScript