Action-level tests for admin access control. green
Following the practice from Section 10.2.1: Requiring logged-in users, we’ll put action-level tests of access control in the Users controller test file. As with the logout test in commit: A test for user logout, we’ll use delete to issue a DELETE request directly to the destroy action. We need to check two cases: first, users who aren’t logged in should be redirected to the login page; second, users who are logged in but who aren’t admins should be redirected to the Home page. The result appears in this commit.

Note that this commit also makes sure that the user count doesn’t change using the assert_no_difference method (seen before in Listing 7.23).

The tests in this commit verify the behavior in the case of an unauthorized (non-admin) user, but we also want to check that an admin can use a delete link to successfully destroy a user. Since the delete links appear on the users index, we’ll add these tests to the users index test from Listing 10.48. The only really tricky part is verifying that a user gets deleted when an admin clicks on a delete link, which we’ll accomplish as follows:

```ruby
assert_difference 'User.count', -1 do
  delete user_path(@other_user)
end
```

This uses the assert_difference method first seen in commit: A test for a valid signup when creating a user, this time verifying that a user is destroyed by checking that User.count changes by −1 when issuing a delete request to the corresponding user path.