An integration test for delete links and destroying users. green
Putting everything together gives the pagination and delete test in this commit, which includes tests for both admins and non-admins.

Note that in this commit: User delete links (viewable only by admins) checks for the right delete links, including skipping the test if the user happens to be the admin (which lacks a delete link due to commit: ).

At this point, our deletion code is well-tested, and the test suite should be green:

$ rails test

Exercises -

By commenting out the admin user before filter in commit: A before filter restricting the destroy action to admins, confirm that the tests go red.

FYI,
modified the users_index_test.rb to make it pass but need to go back and figure out why its not got the else block and not returning false from the validation within the model.