Adding a route for the Account Activations edit action.
At present, newly registered users immediately have full access to their
accounts (Chapter 7: Sign up); in this chapter, we’ll implement an
account activation step to verify that the user controls the email
address they used to sign up. This will involve associating an
activation token and digest with a user, sending the user an email with
a link including the token, and activating the user upon clicking the
link. In Chapter 12: Password reset, we’ll apply similar ideas to allow
users to reset their passwords if they forget them. Each of these two
features will involve creating a new resource, thereby giving us a
chance to see further examples of controllers, routing, and database
migrations. In the process, we’ll also have a chance to learn how to
send email in Rails, both in development and in production.

Our strategy for handling account activation parallels user login
(Section 8.2: Logging in) and especially remembering users (Section 9.1:
Remember me). The basic sequence appears as follows:

  1. Start users in an “unactivated” state.
  2. When a user signs up, generate an activation token and
corresponding activation digest.
  3. Save the activation digest to the database, and then send an email
to the user with a link containing the activation token and user’s email
address.
  4. When the user clicks the link, find the user by email address, and
then authenticate the token by comparing with the activation digest.
  5. If the user is authenticated, change the status from “unactivated”
to “activated”.

Because of the similarity with passwords and remember tokens, we will be
able to reuse many of the same ideas for account activation (as well as
password reset), including the User.digest and User.new_token methods
and a modified version of the user.authenticated? method. Table 11.1
illustrates the analogy (including the password reset from Chapter 12:
Password reset).

As usual, we’ll make a topic branch for the new feature:

```bash
$ git checkout -b account-activation
```

As with Users and Sessions, the actions (or, in this case, the sole
action) for the Account Activations resource will live inside an Account
Activations controller, which we can generate as follows:

```bash
$ rails generate controller AccountActivations
```
As we’ll see in Section 11.2.1: Mailer templates, the activation email
will involve a URL of the form

edit_account_activation_url(activation_token, ...)
which means we’ll need a named route for the edit action. We can arrange
for this with the resources line shown in this commit, which gives the
RESTful route shown in Table 11.2.

We’ll define the edit action itself in Section 11.3.2: Activation edit
action, after we’ve finished the Account Activations data model and
mailers.

Exercises -

1. Verify that the test suite is still green.

Yes, the test suite is still green. After issuing: rails test

2. Why does Table 11.2 list the _url form of the named route instead of
the _path form? Hint: We’re going to use it in an email.

In this case, we've added a the resources route in the routes file not a
single action verb that is normally mapped to a single resource action
path. We've also defined a single edit request action to be associated
with this url in which passed in an id parameter that can return a
single record to edit on.