A migration for account activation (with added index).
As discussed in the introduction, we need a unique activation token for
use in the activation email. One possibility would be to use a string
that’s stored both in the database and included in the activation URL,
but this raises security concerns if our database is compromised. For
example, an attacker with access to the database could immediately
activate newly created accounts (thereby logging in as the user), and
could then change the password to gain control.

To prevent such scenarios, we’ll follow the example of passwords
(Chapter 6: Modeling users) and remember tokens (Chapter 9: Advanced
login) by pairing a publicly exposed virtual attribute with a secure
hash digest saved to the database. This way we can access the activation
token using

```
user.activation_token
```

and authenticate the user with code like

```
user.authenticated?(:activation, token)
```
(This will require a modification of the authenticated? method defined
in commit: Adding an authenticated? method to the User model.)

We’ll also add a boolean attribute called activated to the User model,
which will allow us to test if a user is activated using the same kind
of auto-generated boolean method we saw in Section 10.4.1:

```
if user.activated? ...
```

Finally, although we won’t use it in this tutorial, we’ll record the
time and date of the activation in case we want it for future reference.
The full data model appears in Figure 11.1.

user_model_account_activation

Figure 11.1: The User model with added account activation attributes.

The migration to add the data model from Figure 11.1 adds all three
attributes at the command line:

```bash
$ rails generate migration add_activation_to_users \
> activation_digest:string activated:boolean activated_at:datetime
```

(As noted in Section 8.2.4: Testing layout changes, the > on the second
line is a “line continuation” character inserted automatically by the
shell, and should not be typed literally.) As with the admin attribute
(Listing 10.54), we’ll add a default boolean value of false to the
activated attribute, as shown in this commit.

We then apply the migration as usual:

```bash
$ rails db:migrate
```