Adding account activation code to the User model. green
Activation token callback

Because every newly signed-up user will require activation, we should
assign an activation token and digest to each user object before it’s
created. We saw a similar idea in Section 6.2.5: Uniqueness validation,
where we needed to convert an email address to lower-case before saving
a user to the database. In that case, we used a before_save callback
combined with the downcase method (in commit: Ensuring email uniqueness
by downcasing the email attribute). A before_save callback is
automatically called before the object is saved, which includes both
object creation and updates, but in the case of the activation digest we
only want the callback to fire when the user is created. This requires a
before_create callback, which we’ll define as follows:

```
before_create :create_activation_digest
```

This code, called a method reference, arranges for Rails to look for a
method called create_activation_digest and run it before creating the
user. (In commit: Ensuring email uniqueness by downcasing the email
attribute , we passed before_save an explicit block, but the method
reference technique is generally preferred.) Because the
create_activation_digest method itself is only used internally by the
User model, there’s no need to expose it to outside users; as we saw in
Section 7.3.2, the Ruby way to accomplish this is to use the private
keyword:
```
private

  def create_activation_digest
    # Create the token and digest.
  end
```

All methods defined in a class after private are automatically hidden,
as seen in this console session:

```bash
$ rails console
>> User.first.create_activation_digest
NoMethodError: private method `create_activation_digest' called for
#<User>
```

The purpose of the before_create callback is to assign the token and
corresponding digest, which we can accomplish as follows:

```
self.activation_token  = User.new_token
self.activation_digest = User.digest(activation_token)
```

This code simply reuses the token and digest methods used for the
remember token, as we can see by comparing it with the remember method
from commit-Adding a remember method to the User model:

```

def remember
  self.remember_token = User.new_token
  update_attribute(:remember_digest, User.digest(remember_token))
end
```

The main difference is the use of update_attribute in the latter case.

The reason for the difference is that remember tokens and digests are
created for users that already exist in the database, whereas the
before_create callback happens before the user has been created, so
there’s not yet any attribute to update. As a result of the callback,
when a new user is defined with User.new (as in user signup, in commit:
Using strong parameters in the create action), it will automatically get
both activation_token and activation_digest attributes; because the
latter is associated with a column in the database (Figure 11.1), it
will be written to the database automatically when the user is saved.

Putting together the discussion above yields the User model shown in
this commit. As required by the virtual nature of the activation token,
we’ve added a second attr_accessor to our model. Note that we’ve taken
the opportunity to replace the email downcasing callback from commit:
Ensuring email uniqueness by downcasing the email attribute with a
method reference.