Activating seed users by default + Activating fixture users.
Before moving on, we should also update our seed data and fixtures so
that our sample and test users are initially activated, as shown in this
commit (The Time.zone.now method is a built-in Rails helper that returns
the current timestamp, taking into account the time zone on the server.)

To apply the changes in this commit, reset the database to reseed the
data as usual:

```bash
$ rails db:migrate:reset
$ rails db:seed
```

Exercises -

1. Verify that the test suite is still green after the changes made in
this section.

Yes, the test suite is still green, after issuing: rails test

2. By instantiating a User object in the console, confirm that calling
the create_activation_digest method raises a NoMethodError due to its
being a private method. What is the value of the user’s activation
digest?

Yes, I do receive the following error:

> User.first.create_activation_digest
User.first.create_activation_digest
  User Load (0.5ms)  SELECT  "users".* FROM "users" ORDER BY
"users"."id" ASC LIMIT ?  [["LIMIT", 1]]

        from (irb):1

For the activation digest, we can try calling the following

User.activation_digest
User.activation_digest

        from (irb):2

because the User.activation_digest is known until the User is created
and the create_activation_digest method called that first call

  self.activation_token  = User.new_token

then set the activation_digest in the following code:

  self.activation_digest = User.digest(activation_token)

3. In the commit: An alternate callback implementation, we saw that
email downcasing can be written more simply as email.downcase! (without
any assignment). Make this change to the downcase_email method in this
commit and verify by running the test suite that it works.

Yes, this change has been made.