Generating the User mailer.
Account activation emails

With the data modeling complete, we’re now ready to add the code needed
to send an account activation email. The method is to add a User mailer
using the Action Mailer library, which we’ll use in the Users controller
create action to send an email with an activation link. Mailers are
structured much like controller actions, with email templates defined as
views. These templates will include links with the activation token and
email address associated with the account to be activated.

As with models and controllers, we can generate a mailer using rails
generate, as shown in this commit.

```
$ rails generate mailer UserMailer account_activation password_reset
```

 The command above, generates two view templates for each mailer, one for plain-text email and one for HTML email. For the account activation mailer method, they appear as in

app/views/user_mailer/account_activation.html.erb
app/views/user_mailer/account_activation.text.erb

(We’ll take care of the corresponding password reset templates in git branch: password-reset )