Account activation text & HTML views plus escaped email with CGI.escape
Let’s take a look at the generated mailers code to get a sense of how
they work

Listing 11.9 : app/mailers/application_mailer.rb
Listing 11.10: app/mailers/user_mailer.rb
Listing 11.11: app/mailers/application_mailer.rb
Listing 11.12: app/mailers/user_mailer.rb
Listing 11.13: app/views/user_mailer/account_activation.text.erb
Listing 11.14: app/views/user_mailer/account_activation.html.erb

We see in Listing 11.9 that there is a default from address common to
all mailers in the application, and each method in Listing 11.10 has a
recipient’s address as well. (Listing 11.9 also uses a mailer layout
corresponding to the email format; although it won’t ever matter in this
tutorial, the resulting HTML and plain-text mailer layouts can be found
in app/views/layouts.) The generated code also includes an instance
variable (@greeting), which is available in the mailer views in much the
same way that instance variables in controllers are available in
ordinary views.

Let’s take a look at the generated mailers to get a sense of how they
work (Listing 11.9 and Listing 11.10). We see in Listing 11.9 that there
is a default from address common to all mailers in the application, and
each method in Listing 11.10 has a recipient’s address as well. (Listing
11.9 also uses a mailer layout corresponding to the email format;
although it won’t ever matter in this tutorial, the resulting HTML and
plain-text mailer layouts can be found in app/views/layouts.) The
generated code also includes an instance variable (@greeting), which is
available in the mailer views in much the same way that instance
variables in controllers are available in ordinary views.

To make a working activation email, we’ll first customize the generated
template as shown in Listing 11.11. Next, we’ll create an instance
variable containing the user (for use in the view), and then mail the
result to user.email (Listing 11.12). As seen in Listing 11.12, the mail
method also takes a subject key, whose value is used as the email’s
subject line.

As with ordinary views, we can use embedded Ruby to customize the
template views, in this case greeting the user by name and including a
link to a custom activation link. Our plan is to find the user by email
address and then authenticate the activation token, so the link needs to
include both the email and the token. Because we’re modeling activations
using an Account Activations resource, the token itself can appear as
the argument of the named route defined in Listing 11.1:

```
edit_account_activation_url(@user.activation_token, ...)
```

Recalling that

```
edit_user_url(user)
```

produces a URL of the form

http://www.example.com/users/1/edit
the corresponding account activation link’s base URL will look like
this:

```
http://www.example.com/account_activations/q5lt38hQDc_959PVoo6b7A/edit
```
Here q5lt38hQDc_959PVoo6b7A is a URL-safe base64 string generated by the
new_token method (Listing 9.2), and it plays the same role as the user
id in /users/1/edit. In particular, in the Activations controller edit
action, the token will be available in the params hash as params[:id].

In order to include the email as well, we need to use a query parameter,
which in a URL appears as a key-value pair located after a question
mark:6

```
account_activations/q5lt38hQDc_959PVoo6b7A/edit?email=foo%40example.com
```

Notice that the ‘@’ in the email address appears as %40, i.e., it’s
“escaped out” to guarantee a valid URL. The way to set a query parameter
in Rails is to include a hash in the named route:

```
edit_account_activation_url(@user.activation_token, email: @user.email)
```
When using named routes in this way to define query parameters, Rails
automatically escapes out any special characters. The resulting email
address will also be unescaped automatically in the controller, and will
be available via params[:email].

With the @user instance variable as defined in Listing 11.12, we can
create the necessary links using the named edit route and embedded Ruby,
as shown in Listing 11.13 and Listing 11.14. Note that the HTML template
in Listing 11.14 uses the link_to method to construct a valid link.

Exercises -

At the console, verify that the escape method in the CGI module escapes
out the email address as shown in Listing 11.15. What is the escaped
value of the string "Don’t panic!"?
Here is the verification for the ecsaped of  'foo@example.com'

CGI.escape('foo@example.com')
=> "foo%40example.com"

Here is the verification for the ecsaped of  'Don’t panic!'

CGI.escape('Don’t panic!')
=> "Don%E2%80%99t+panic%21"