Using the generalized authenticated? method in current_user. red
The reason for the failure is that the current_user method (commit: A
test of authenticated? with a nonexistent digest) and the test for nil
digests (commit: A test of authenticated? with a nonexistent digest)
both use the old version of authenticated?, which expects one argument
instead of two. To fix this, we simply update the two cases to use the
generalized method, as shown in Listing 11.28 and Listing 11.29.