An edit action to activate accounts.
 With the authenticated? method as in commit: A generalized
authenticated? method, we’re now ready to write an edit action that
authenticates the user corresponding to the email address in the params
hash. Our test for validity will look like this:

```
if user && !user.activated? && user.authenticated?(:activation,
params[:id])
```

Note the presence of !user.activated?, which is the extra boolean
alluded to above. This prevents our code from activating users who have
already been activated, which is important because we’ll be logging in
users upon confirmation, and we don’t want to allow attackers who manage
to obtain the activation link to log in as the user.

If the user is authenticated according to the booleans above, we need to
activate the user and update the activated_at timestamp:

```
user.update_attribute(:activated,    true)
user.update_attribute(:activated_at, Time.zone.now)
```

This leads to the edit action shown in this commit. Note also that this
commit handles the case of an invalid activation token; this should
rarely happen, but it’s easy enough to redirect in this case to the root
URL.

With the code in this commit, you should now be able to paste in the URL
from the commit that showed a sample account activation email from the
server log to activate the relevant user. For example, on my system I
visited the URL

https://localhost:3000/account_activations/
$2a$10$ZLb.A9Smrk3vPbNmO3q6s.LisPFZhp6xCrOksFXLvJWe94qcxKRRO/edit?email=user1@example.com

and got the result of a