Revised: Using the generalized authenticated? method in the User test
Discovered that the issue with provided authenticated? method and had
was an issue but the inputed token was not correct. In fact, the token
was not the right one, when I ran the activation generated one:

http://localhost:3000/account_activations/Mhcve7dF9Y7ZLJxm3Xrw1g/edit?email=tosmalls%40gmail.com

What much shorter and created with the following private method:

create_activation_digest

Its difficult to reproduce but this something keep in mind when trying
to uses the authenticated as coded. I might include the previous commit
code in the future.