Adding account activation to the user signup test. green
In this section, we’ll add an integration test for account activation.
Because we already have a test for signing up with valid information,
we’ll add the steps to the test developed in Section 7.4.4: A test for
valid submission (commit: A test for a valid signup.). There are quite a
few steps, but they are mostly straightforward; see if you can follow
along in this commit. (The highlights in this commit indicate lines that
are especially important or easy to miss, but there are other new lines
as well, so take care to add them all.)

There’s a lot of code in this commit, but the only completely novel code
is in the line

```
assert_equal 1, ActionMailer::Base.deliveries.size
```

This code verifies that exactly 1 message was delivered. Because the
deliveries array is global, we have to reset it in the setup method to
prevent our code from breaking if any other tests deliver email (as will
be the case in Chapter 12). In this commut, also uses the assigns method
for the first time in the main tutorial; as explained in a Chapter 9
exercise (Section 9.3.1.1: A test of the “remember me” checkbox),
assigns lets us access instance variables in the corresponding action.
For example, the Users controller’s create action defines an @user
variable (commit: Adding account activation to user signup), so we can
access it in the test using assigns(:user). Finally, note that this
commit restores the lines we commented out in commit: Temporarily
commenting out failing tests.

At this point, the test suite should be green:

$ rails test

Note: that I've commented out line# 35 because it caused line# 36 to
fail. Because it will return that the user is already logged in.