Adding user activation methods to the User model.
With the test in commit: Adding account activation to the user signup
test., we’re ready to refactor a little by moving some of the user
manipulation out of the controller and into the model. In particular,
we’ll make an activate method to update the user’s activation attributes
and a send_activation_email to send the activation email. The extra
methods appear in Listing 11.35, and the refactored application code
appears in Listing 11.36 and Listing 11.37.

Listing 11.35: Adding user activation methods to the User model :
app/models/user.rb

Listing 11.36: Sending email via the user model object.
app/controllers/users_controller.rb

Listing 11.37: Account activation via the user model object.
app/controllers/account_activations_controller.rb

Note that Listing 11.35 eliminates the use of user., which would break
inside the User model because there is no such variable:

```
-user.update_attribute(:activated,    true)
-user.update_attribute(:activated_at, Time.zone.now)
+update_attribute(:activated,    true)
+update_attribute(:activated_at, Time.zone.now)
```

(We could have switched from user to self, but recall from Section 6.2.5
that self is optional inside the model.) It also changes @user to self
in the call to the User mailer:

```
-UserMailer.account_activation(@user).deliver_now
+UserMailer.account_activation(self).deliver_now
```

These are exactly the kinds of details that are easy to miss during even
a simple refactoring but will be caught by a good test suite. Speaking
of which, the test suite should still be green:

$ rails test
Started with run options --seed 48197

Progress: |Run options: --seed 48197

=.=.==.==.=.=..====.=..====..===..===...====..====...=====..===.=.==..===.=.==.==.=.==.==.=.==.=.==.==.=..====|
.
Finished in 17.10466s
43 tests, 193 assertions, 0 failures, 0 errors, 0 skips

Finished in 17.312980s, 2.4837 runs/s, 11.1477 assertions/s.
43 runs, 193 assertions, 0 failures, 0 errors, 0 skips

1. In Listing 11.35, the activate method makes two calls to the
update_attribute, each of which requires a separate database
transaction. By filling in the template shown in Listing 11.39, replace
the two update_attribute calls with a single call to update_columns,
which hits the database only once. After making the changes, verify that
the test suite is still green.

2. Right now all users are displayed on the user index page at /users
and are visible via the URL /users/:id, but it makes sense to show users
only if they are activated. Arrange for this behavior by filling in the
template shown in Listing 11.40.9 (This uses the Active Record where
method, which we’ll learn more about in Section 13.3.3.)

Couldn't the secord portion to work; I'll look at it later.

3. To test the code in the previous exercise, write integration tests
for both /users and /users/:id.

Moving will revite later.