Adding a resource for password resets.
Having completed account activation (and thereby verified the user’s
email address) in Chapter 11, we’re now in a good position to implement
password reset, and thereby handle the common case of users forgetting
their passwords. As we’ll see, many of the steps are similar to those
for account activation, and we will have several opportunities to apply
the lessons learned in Chapter 11. The beginning is different, though;
unlike account activation, implementing password resets requires both a
change to one of our views and two new forms (to handle email and new
password submission).

Before writing any code, let’s mock up the expected sequence for
resetting passwords. We’ll start by adding a “forgot password” link to
the sample application’s login form. The “forgot password” link will go
to a page with a form that takes in an email address and sends an email
containing a password reset link. The reset link will go to a form for
resetting the user’s password (with confirmation).

If you followed Chapter 11: Account activation, you already have a
mailer for password resets, which was generated in Section 11.2: Account
activation emails (commit:  Generating the User mailer). In this
section, we’ll complete the necessary preliminaries by adding a resource
and data model for password resets (Section 12.1) to go along with the
mailer. We’ll implement the actual password reset in Section 12.3.

In analogy with account activations, our general plan is to make a
Password Resets resource, with each password reset consisting of a reset
token and corresponding reset digest. The primary sequence goes like
this:

1. When a user requests a password reset, find the user by the submitted
email address.
2. If the email address exists in the database, generate a reset token
and corresponding reset digest.
3. Save the reset digest to the database, and then send an email to the
user with a link containing the reset token and user’s email address.
4. When the user clicks the link, find the user by email address, and
then authenticate the token by comparing it to the reset digest.
5. If authenticated, present the user with the form for changing the
password.

As with sessions (Section 8.1: Sessions) and account activations
(Chapter 11: Account activation), we’ll model password resets as a
resource even though they won’t be associated with an Active Record
model. Instead, we’ll include the relevant data (including the reset
token) in the User model itself.

Because we’ll be treating password resets as a resource, we’ll interact
with them via the standard REST URLs. Unlike the activation link, which
required only an edit action, in this case we’ll be rendering both new
and edit forms for manipulating password resets, as well as creating and
updating them, so we’ll end up using four RESTful routes in total.

As usual, we’ll make a topic branch for the new feature:

```
$ git checkout -b password-reset
```

Our first step is to generate a controller for the Password Resets
resource, in this case making both new and edit actions per the
discussion above:

```
$ rails generate controller PasswordResets new edit --no-test-framework
```

Note that we’ve included a flag to skip generating tests. This is
because we don’t need the controller tests, preferring instead to build
on the integration test from Section 11.3.3: Activation test and
refactoring.

Because we’ll need forms both for creating new password resets and for
updating them by changing the password in the User model, we need routes
for new, create, edit, and update. We can arrange this with the
resources line shown in this commit.

The code in this commit arranges for the RESTful routes shown in Table 12.1. In particular, the first route in Table 12.1 gives a link to the “forgot password” form via