A new password reset view.
To make the view for new password resets, we’ll work in analogy with the
previous form for making a new non–Active Record resource, namely, the
login form (commit: Code for the login form) for creating a new session,
shown again in

Listing 12.3: Reviewing the code for the login form.
app/views/sessions/new.html.erb

for reference.

The new password resets form has a lot in common with Listing 12.3; the
most important differences are the use of a different resource and URL
in the call to form_for and the omission of the password attribute. The
result appears in this commit.

Exercises -

1. Why does the form_for in this commit use :password_reset instead of
@password_reset?

:password_reset is a symbol acts like a container instead of a instance
variable. That contrainer is converted to array that holds all the form
fields in the generated html.