A create action for password resets.
Upon submitting the new form, we need to find the user by email address
and update its attributes with the password reset token and sent-at
timestamp. We then redirect to the root URL with an informative flash
message. As with login (Listing 8.11), in the case of an invalid
submission we re-render the new page with a flash.now message. The
results appear in this commit.