The form to reset a password.
Now that we have a correctly generated email as in the commit: A sample
password reset email from the server log, we need to write the edit
action in the Password Resets controller that actually resets the user’s
password. As in Section 11.3.3: Activation test and refactoring, we’ll
write a thorough integration test as well.

Password reset emails such as that shown in commit: A sample password
reset email from the server log, contain links of the following form:

```
https://example.com/password_resets/3BdBrXeQZSWqFIDRN8cxHA/edit?email=fu%40bar.com
```

To get these links to work, we need a form for resetting passwords. The
task is similar to updating users via the user edit view (commit: The
user edit view), but in this case with only password and confirmation
fields.

There’s an additional complication, though: we expect to find the user
by email address, which means we need its value in both the edit and
update actions. The email will automatically be available in the edit
action because of its presence in the link above, but after we submit
the form its value will be lost. The solution is to use a hidden field
to place (but not display) the email on the page, and then submit it
along with the rest of the form’s information. The result appears in
this commit.

Note that in this commit uses the form tag helper

```
hidden_field_tag :email, @user.email
```
instead of

```
f.hidden_field :email, @user.email
```
because the reset link puts the email in params[:email], whereas the
latter would put it in params[:user][:email].