The edit action for password reset
To get the form to render, we need to define an @user variable in the
Password Resets controller’s edit action. As with account activation
(Listing 11.31), this involves finding the user corresponding to the
email address in params[:email]. We then need to verify that the user is
valid, i.e., that it exists, is activated, and is authenticated
according to the reset token from params[:id] (using the generalized
authenticated? method defined in commit: A generalized authenticated?
method). Because the existence of a valid @user is needed in both the
edit and update actions, we’ll put the code to find and validate it in a
couple of before filters, as shown in this commit.

In commit: The edit action for password reset, compare the use of

```
authenticated?(:reset, params[:id])
```

to

```
authenticated?(:remember, cookies[:remember_token])
```
in commit: Using the generalized authenticated? method in current_user
and

authenticated?(:activation, params[:id])
in commit: An edit action to activate accounts Together, these three
uses complete the authentication methods shown in Table 11.1.

With the code as above, following the link from Listing 12.11 should
render a password reset form.

Exercises -

1. Follow the link in the email from the server log in Section 12.2.1.1.
Does it properly render the form as shown in Figure 12.11?

Yes it does and I've submitted a new password: foobar

2. What happens if you submit the form from the previous exercise?

Error:

Started PATCH "/password_resets/42t5E3XB2hVrYF_ZFsmCwQ" for 127.0.0.1 at
2017-12-29 22:49:35 -0500

AbstractController::ActionNotFound (The action 'update' could not be
found for PasswordResetsController):