The update action for password reset.
Unlike the Account Activations edit method, which simply toggles the
user from “inactive” to “active”, the edit method for Password Resets is
a form, which must therefore submit to a corresponding update action. To
define this update action, we need to consider four cases:

  1. An expired password reset
  2. A failed update due to an invalid password
  3. A failed update (which initially looks “successful”) due to an
empty password and confirmation
  4. A successful update

Cases (1), (2), and (4) are fairly straightforward, but Case (3) is
non-obvious and is explained in more detail below.

Case (1) applies to both the edit and update actions, and so logically
belongs in a before filter:

```
before_action :check_expiration, only: [:edit, :update]    # Case (1)
```

This requires defining a private check_expiration method:

def check_expiration
  if @user.password_reset_expired?
    flash[:danger] = "Password reset has expired."
    redirect_to new_password_reset_url
  end
end

In the check_expiration method, we’ve deferred the expiration check to
the instance method password_reset_expired?, which is a little tricky
and will be defined in a moment.

Commit: The update action for password reset shows the implementation of
these filters, together with the update action that implements Cases
(2)–(4). Case (2) gets handled by a failed update, with the error
messages from the shared partial in commit: The form to reset a password
displaying automatically when the edit form is re-rendered. Case (4)
corresponds to a successful change, and the result is similar to a
successful login (commit: Logging in the user upon signup).

The only failure case not handled by Case (2) is when the password is
empty, which is currently allowed by our User model (Listing 10.13) and
so needs to be caught and handled explicitly. This is Case (3) above.
Our method in this case is to add an error directly to the @user
object’s error messages using errors.add:

```
@user.errors.add(:password, :blank)
```

This arranges to use the default message for blank content when the
password is empty.

The result of putting Cases (1)–(4) together is the update action shown
in commit:  The update action for password reset.